backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-filter
drf-url-filters==0.5.1
drf-spectacular==0.27.2
psycopg2-binary>=2.9.10
gunicorn>=21.0.0
django-environ>=0.11.0
python-dotenv>=1.0.0
Pillow>=10.1.0
requests>=2.26.0